feat(codegen): preserve code coverage ignore comments (e.g. `v8 ignore`) (#10338)
fixes #10091

This PR attempts to preserve any comments with a start match of

```
/* v8 ignore */
/* c8 ignore */
/* node:coverage */
/* istanbul ignore */
```

Added a special handle for coverage comment before `catch`:

```
try {
  something();
}
/* istanbul ignore next */
catch(e) {
  // should never happen
}
```
